+++ R/ropenblas.R
@@ -94,9 +94,10 @@ exist <- function(x = "gcc") {
   ifelse(length(result) == 1L, FALSE, TRUE)
 }
 
+#' @importFrom cli style_bold col_red
 validate_answer <- function(x) {
   if (!(x %in% c("y", "no", "yes", "n")))
-    stop("Invalid option. Procedure interrupted.")
+    stop(glue("[{style_bold(col_red(symbol$cross))}] Invalid option. Procedure interrupted."))
 }
 
 modern_openblas <- function(x) {
@@ -147,6 +148,22 @@ loop_root <- function(x, attempt = 3L, sudo = TRUE) {
 answer_yes_no <- function(text) {
   readline(prompt = glue("{text} (yes/no): ")) %>%
     tolower
+}
+
+#' @importFrom glue glue
+#' @importFrom magrittr "%>%"
+#' @importFrom stringr str_detect
+error_r <- function(){
+  #current_directory <- "{R.home()}/bin" %>% glue
+  run_r <- "cd /usr/bin && ./R --no-save&" %>% 
+    glue %>% 
+    system(intern = TRUE)
+  
+  error <- any(FALSE, str_detect(run_r, pattern = "Error:"))
+  
+  if(length(error) == 0L) error <- TRUE
+  
+  error
 }
 
 #' @title Download, Compile and Link OpenBLAS Library with \R
@@ -203,22 +220,25 @@ answer_yes_no <- function(text) {
 #' @export
 ropenblas <- function(x = NULL, restart_r = TRUE) {
   if (Sys.info()[[1L]] != "Linux")
-    stop("Sorry, this package for now configures R to use the OpenBLAS library on Linux systems.\n")
+    stop("Sorry, this package for now configures {style_bold(\"R\")} to use the {style_bold(\"OpenBLAS\")} library on Linux systems.\n")
   
   if (!connection())
-    stop("You apparently have no internet connection.\n")
+    stop(glue("[{style_bold(col_red(symbol$cross))}] You apparently have no internet connection.\n"))
   
   download <- download_openblas(x)
   repo <- download$repo_openblas
   diretory_tmp <- download$path_openblas
   
+  "cp {dir_blas()$path}{dir_blas()$file_blas} {dir_blas()$path}{dir_blas()$file_blas}.keep" %>%
+    glue %>% 
+    loop_root(attempt = 5L)
+  
   if (!is.null(x) && glue("v{x}") > download$version)
     stop(
       glue(
-        "Version {x} does not exist. The latest version of OpenBLAS is {substr(download$version, 2L, nchar(download$version))}."
+        "{symbol$bullet} Version {style_bold({x})} does not exist. The latest version of {style_bold(\"OpenBLAS\")} is {style_bold({substr(download$version, 2L, nchar(download$version))})}."
       )
     )
-  
   
   if (!is.null(x)) {
     if (dir_blas()$use_openblas) {
@@ -226,7 +246,7 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
         cat("\n")
         if (glue("v{x}") != download$version) {
           answer <-
-            "The latest version of OpenBLAS is {substr(download$version, 2L, nchar(download$version))}. Do you want to install this version?" %>%
+            "{symbol$bullet} The latest version of {style_bold(\"OpenBLAS\")} is {style_bold({{substr(download$version, 2L, nchar(download$version))})}. Do you want to install this version?" %>%
             glue %>%
             answer_yes_no
           
@@ -244,7 +264,8 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
       } else {
         if (glue("v{dir_blas()$version_openblas}") == download$version) {
           answer <-
-            "The latest version of OpenBLAS is already in use. Do you want to compile and link again?" %>%
+            "{symbol$bullet} The latest version of {style_bold(\"OpenBLAS\")} is already in use. Do you want to compile and link again?" %>%
+            glue %>% 
             answer_yes_no
           
           validate_answer(answer)
@@ -257,7 +278,7 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
         } else {
           stop(
             glue(
-              "There is no OpenBLAS version {x}. The latest version is {substr(download$version, 2L, nchar(download$version))}."
+              "{symbol$bullet} There is no {style_bold(\"OpenBLAS\")} version {style_bold({x})}. The latest version is {style_bold({{substr(download$version, 2L, nchar(download$version))}})}."
             )
           )
         }
@@ -265,7 +286,7 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
     } else {
       if (glue("v{x}") < download$version) {
         answer <-
-          "The latest version is {substr(download$version, 2L, nchar(download$version))}. Want to consider the latest version?" %>%
+          "{symbol$bullet} The latest version is {style_bold({{substr(download$version, 2L, nchar(download$version))}})}. Want to consider the latest version?" %>%
           glue %>%
           answer_yes_no
         
@@ -288,7 +309,8 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
         checkout(repo, download$version)
       } else {
         answer <-
-          "The latest version of OpenBLAS is already in use. Do you want to compile and link again?" %>%
+          "{symbol$bullet} The latest version of {style_bold(\"OpenBLAS\")} is already in use. Do you want to compile and link again?" %>%
+          glue %>% 
           answer_yes_no
         validate_answer(answer)
         
@@ -317,10 +339,10 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
   
   if (!exist())
     stop(
-      "GNU GCC not installed. Install GNU GCC Compiler (C and Fortran) on your operating system."
+      glue("{style_bold(col_red(symbol$cross))} GNU GCC not installed. Install GNU GCC Compiler (C and Fortran) on your operating system.")
     )
   if (!exist("make"))
-    stop("GNU Make not installed. Install GNU Make on your operating system.")
+    stop(glue("{style_bold(col_red(symbol$cross))} GNU Make not installed. Install GNU Make on your operating system."))
   
   if (!exist_opt())
     mkdir_opt()
@@ -331,15 +353,39 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
     diretory_tmp
   }) %>% setwd
   
-  glue("sudo -kS make install PREFIX=/opt/OpenBLAS") %>%
+  glue("make install PREFIX=/opt/OpenBLAS") %>%
     loop_root(attempt = 5L)
   
   setwd(dir_blas()$path)
   
   if (!str_detect(dir_blas()$file_blas, "libopenblas")) {
     glue(
-      "sudo -kS ln -snf /opt/OpenBLAS/lib/libopenblas.so {dir_blas()$path}{dir_blas()$file_blas}"
+      "ln -snf /opt/OpenBLAS/lib/libopenblas.so {dir_blas()$path}{dir_blas()$file_blas}"
     ) %>% loop_root(attempt = 5L)
+  }
+  
+  if (error_r()) {
+    "mv {dir_blas()$path}{dir_blas()$file_blas}.keep {dir_blas()$path}{dir_blas()$file_blas}" %>% 
+      glue %>% 
+      loop_root(attempt = 5L)
+  
+    cat("\n")
+    
+    cat(
+      rule(
+        width = 50L,
+        center = glue("{style_bold(\"Procedure Incompleted\")}"),
+        col = "red",
+        background_col = "gray90",
+        line = 2L
+      )
+    )    
+    
+    "[{style_bold(symbol$cross)}] Some error has occurred. No changes have been made." %>% 
+      glue %>% 
+      warning %>% 
+      return
+    
   }
   
   .refresh_terminal <- function() {
@@ -358,9 +404,9 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
   cat("\n")
   
   cat(
-    cli::rule(
-      width = 35L,
-      center = "Procedure Completed",
+    rule(
+      width = 50L,
+      center = glue("{style_bold(\"Procedure Completed\")}"),
       col = "blue",
       background_col = "gray90",
       line = 2L
@@ -370,12 +416,12 @@ ropenblas <- function(x = NULL, restart_r = TRUE) {
   cat("\n")
   
   if (is.null(x)) {
-    "[{style_bold(col_green(symbol$tick))}] OpenBLAS version {substr(download$version, 2L, nchar(download$version))}." %>%
+    "[{style_bold(col_green(symbol$tick))}] {style_bold(\"OpenBLAS\")} version {style_bold({{substr(download$version, 2L, nchar(download$version))}})}." %>%
       glue %>%
       cat
     
   } else {
-    "[{style_bold(col_green(symbol$tick))}] OpenBLAS version {x}." %>%
+    "[{style_bold(col_green(symbol$tick))}] {style_bold(\"OpenBLAS\")} version {style_bold({x})}." %>%
       glue %>%
       cat
   }
@@ -514,9 +560,9 @@ attention <- function(x) {
   cat("\n")
   
   cat(
-    cli::rule(
+    rule(
       width = 60L,
-      center = "VERY ATTENTION",
+      center = glue("{style_bold(\"VERY ATTENTION\")}"), 
       col = "red",
       background_col = "gray90",
       line = 2L
@@ -528,7 +574,7 @@ attention <- function(x) {
   glue(
     'The {style_bold("ropenblas")} package depends on versions of {style_bold("R")} ',
     '{style_bold(symbol$geq)} {style_bold("3.1.0")}. Therefore,',
-    'you will not be able to \nuse this package to go back to a later version of R!'
+    'you will not be able to \nuse this package to go back to a later version of {style_bold("R")}!'
   ) %>% cat
   
   answer <- NULL
@@ -554,7 +600,7 @@ change_r <- function (x, change = TRUE) {
   dir_r  <- R.home("bin")
   
   if (change) {
-    "sudo -kS ln -sf /opt/R/{x}/lib64/R/bin/R /usr/bin/R"  %>% 
+    "ln -sf /opt/R/{x}/lib64/R/bin/R /usr/bin/R"  %>% 
       glue %>% 
       loop_root(attempt = 5L)
     
@@ -567,8 +613,8 @@ change_r <- function (x, change = TRUE) {
   
   cat(
     rule(
-      width = 35L,
-      center = "Procedure Completed",
+      width = 50L,
+      center = glue("{style_bold(\"Procedure Completed\")}"), 
       col = "blue",
       background_col = "gray90",
       line = 2L
@@ -577,13 +623,13 @@ change_r <- function (x, change = TRUE) {
   
   cat("\n")
   
-  "[{style_bold(col_green(symbol$tick))}] R version {x}." %>%
+  "[{style_bold(col_green(symbol$tick))}] {style_bold(\"R\")} version {style_bold({x})}." %>%
     glue %>%
     cat
   
   cat("\n")
   
-  "{symbol$mustache} The roles are active after terminating the current R session ..." %>%
+  "{symbol$mustache} The roles are active after terminating the current {style_bold(\"R\")} session ..." %>%
     glue %>%
     col_blue %>%
     style_bold %>%
@@ -624,7 +670,7 @@ change_r <- function (x, change = TRUE) {
 rcompiler <- function(x = NULL,
                       version_openblas = NULL) {
   if (Sys.info()[[1L]] != "Linux")
-    stop("Sorry, this package for now configures R to use the OpenBLAS library on Linux systems.\n")
+      stop("Sorry, this package for now configures R to use the OpenBLAS library on Linux systems.\n")
   
   if (!connection())
     stop("You apparently have no internet connection.\n")
@@ -673,10 +719,10 @@ rcompiler <- function(x = NULL,
     glue("make -j $(nproc)") %>%
       system
     
-    glue("sudo -kS make install PREFIX=/opt/R/{x}") %>%
+    glue("make install PREFIX=/opt/R/{x}") %>%
       loop_root(attempt = 5L)
     
-    glue("sudo -kS ln -sf /opt/R/{x}/lib64/R/bin/R /usr/bin/R")  %>%
+    glue("ln -sf /opt/R/{x}/lib64/R/bin/R /usr/bin/R")  %>%
       loop_root(attempt = 5L)
     
     if (!is.null(version_openblas)) {
@@ -686,8 +732,8 @@ rcompiler <- function(x = NULL,
       
       cat(
         rule(
-          width = 35L,
-          center = "Procedure Completed",
+          width = 50L,
+          center = glue("{style_bold(\"Procedure Completed\")}"),
           col = "blue",
           background_col = "gray90",
           line = 2L
@@ -707,11 +753,11 @@ rcompiler <- function(x = NULL,
     glue("make -j $(nproc)") %>%
       system
     
-    glue("sudo -kS make install PREFIX=/opt/R/{x}") %>%
+    glue("make install PREFIX=/opt/R/{x}") %>%
       loop_root(attempt = 5L)
     
     
-    "sudo -kS ln -sf /opt/R/{x}/lib64/R/bin/R /usr/bin/R"  %>% 
+    "ln -sf /opt/R/{x}/lib64/R/bin/R /usr/bin/R"  %>% 
       glue %>% 
       loop_root(attempt = 5L)
     
@@ -725,13 +771,13 @@ rcompiler <- function(x = NULL,
   
   cat("\n")
   
-  "[{style_bold(col_green(symbol$tick))}] R version {x}." %>%
+  "[{style_bold(col_green(symbol$tick))}] {style_bold(\"R\")} version {style_bold({x})}." %>%
     glue %>%
     cat
   
   cat("\n")
   
-  "{symbol$mustache} The roles are active after terminating the current R session ..." %>%
+  "{symbol$mustache} The roles are active after terminating the current {style_bold(\"R\")} session ...\n\n" %>%
     glue %>%
     col_blue %>%
     style_bold %>%