+++ NEWS.md
@@ -1,3 +1,18 @@
+# ropenblas 0.2.2 (development version)
+
+* Code review;
+  
+* Improve code efficiency.
+
+# ropenblas 0.2.1 (latest stable version)
+
+* Code review;
+
+* Highlighting some messages;
+
+* In the tests performed, apparently an error will not break the installed version of R.
+
+
 # ropenblas 0.2.0
 
 * Function `last_version_r(major = 3L)` implemented.  Given the higher version, the function will return the latest stable version of the R language. Major release number of R language (eg. `1L`, `2L`, `3L`, ...);
@@ -21,6 +36,3 @@
 * The `ropenblas()` function documentation describes the details of changes that are made to the system;
 
 * The `ropenblas()` function is capable of suggesting a stable and newer version of the OpenBLAS library. The user can decide whether to use the latest version or the one of his choice. By default if no argument is passed to the `ropenblas()` function, the latest version of the OpenBLAS library will be considered.
-
-
-